+++ lib/CMakeLists.txt
@@ -1,7 +1,7 @@
 project(core VERSION 0.1 LANGUAGES CXX)
 
 
-file(GLOB PROJECT_SOURCES "*.cpp" "*.h" "*.hpp" "*.cc" "*.c" "*.ui")
+file(GLOB PROJECT_SOURCES "*.cpp" "*.h" "*.hpp" "*.cc" "*.c" "*.ui" "data/*.cpp" "data/*.h" "data/*.hpp" "data/*.cc" "data/*.c" "data/*.ui" "plots/*.cpp" "plots/*.h" "plots/*.hpp" "plots/*.cc" "plots/*.c" "plots/*.ui")
 
 if(${QT_VERSION_MAJOR} GREATER_EQUAL 6)
 qt_add_library(core
@@ -36,7 +36,7 @@ MACOSX_BUNDLE_BUNDLE_VERSION ${PROJECT_VERSION}
 MACOSX_BUNDLE_SHORT_VERSION_STRING ${PROJECT_VERSION_MAJOR}.${PROJECT_VERSION_MINOR}
 )
 
-target_sources(core INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/graph_settings_model.h)
+target_sources(core INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/plots/graph_settings_model.h)
 target_include_directories(core INTERFACE ${CMAKE_CURRENT_BINARY_DIR}/core_autogen/include)
 
 if(QT_VERSION_MAJOR EQUAL 6)
