+++ lib/data/manager.cc
