+++ lib/data/variable_data.h
