+++ lib/plots/graph_settings_model.h
