+++ lib/plots/lineplot.ui
@@ -53,7 +53,7 @@
   <customwidget>
    <class>SettingsModel</class>
    <extends>QTableWidget</extends>
-   <header>../../../../lib/graph_settings_model.h</header>
+   <header>../../../../lib/plots/graph_settings_model.h</header>
   </customwidget>
  </customwidgets>
  <resources/>
